requirements.txt
aiobotocore==1.4.1
aiohttp==3.8.1
aiohttp-retry==2.4.6
aioitertools==0.8.0
aiosignal==1.2.0
alembic==1.4.1
anyio==3.3.4
appdirs==1.4.4
argon2-cffi==21.1.0
async-timeout==4.0.1
atpublic==2.3
attrs==19.3.0
Automat==0.8.0
Babel==2.9.1
backcall==0.1.0
bleach==3.1.1
blinker==1.4
boto3==1.19.8
botocore==1.20.106
cached-property==1.5.1
catboost==1.0.3
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.3
cloud-init==22.2
cloudpickle==2.0.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cramjam==2.5.0
cryptography==2.8
cycler==0.11.0
databricks-cli==0.16.2
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.4.0
distro-info===0.23ubuntu1
docker==5.0.3
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.3
entrypoints==0.3
fastparquet==0.8.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flatten-dict==0.4.2
flufl.lock==6.0
fonttools==4.28.1
frozenlist==1.2.0
fsspec==2021.11.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
grandalf==0.6
graphviz==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.8.2
importlib-resources==5.4.0
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.0.1
jedi==0.15.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
json5==0.9.6
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyter-server==1.11.2
jupyterlab==3.2.3
jupyterlab-server==2.8.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightgbm==3.3.1
mailchecker==4.0.16
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.0
mistune==0.8.4
mlflow==1.20.2
more-itertools==4.2.0
multidict==5.2.0
nanotime==0.5.2
nbclassic==0.3.4
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.6.3
notebook==6.0.3
numpy==1.21.4
oauthlib==3.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.4.2
parso==0.5.2
pathspec==0.9.0
pexpect==4.6.0
phonenumbers==8.12.37
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
ply==3.11
prometheus-client==0.7.1
prometheus-flask-exporter==0.18.5
prompt-toolkit==2.0.10
protobuf==3.19.1
psutil==5.8.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydot==1.4.2
pygit2==1.7.0
Pygments==2.10.0
PyGObject==3.36.0
pygtrie==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.34
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-benedict==0.24.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-editor==1.0.4
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.3.1
pyzmq==18.1.1
querystring-parser==1.2.4
requests==2.22.0
requests-unixsocket==0.2.0
rich==10.14.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3fs==2021.11.0
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
shortuuid==1.0.8
shtab==1.5.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sniffio==1.2.0
sos==4.3
SQLAlchemy==1.4.27
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
terminado==0.12.1
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
threadpoolctl==3.0.0
toml==0.10.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
transliterate==1.10.2
Twisted==18.9.0
typing_extensions==4.0.0
ubuntu-advantage-tools==27.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
voluptuous==0.12.2
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==2.0.0
wrapt==1.13.3
xgboost==1.5.0
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
zope.interface==4.7.1
